Fixed cppcheck

+++ src/webots/core/WbSysInfo.cpp
@@ -366,7 +366,7 @@ bool WbSysInfo::isVirtualMachine() {
   __get_cpuid(0x1, &eax, &ebx, &ecx, &edx);
   // cppcheck-suppress shiftTooManyBitsSigned
   // cppcheck-suppress integerOverflow
-  if (ecx & (1 << 31)) {
+  if (ecx & ((unsigned int)1 << 31)) {
     virtualMachine = 1;
     return true;
   }
